requirements.txt
fastapi[all]
pymongo
motor
python-dotenv
bcrypt
pyjwt
uvicorn